+++ util.h
@@ -8,12 +8,12 @@
 // roll is a measure of counter-clockwise rotation around the x axis
 // pitch is the same but around the y axis, and yaw is around the z axis
 typedef struct {
-    double roll, pitch, yaw;
+    float roll, pitch, yaw;
 } rotation_vec3;
 
 // cartesian vector 
 typedef struct {
-    double x, y, z;
+    float x, y, z;
 } vec3;
 
 // multiplies a 3x1 vector over a 3x3 matrix
@@ -26,7 +26,7 @@ vec3 multiply_vector_over_matrix(vec3 matrix[3], vec3 vector);
 // allocate the passed vec3 array on the stack and pass the address to matrix
 // note: rotations are expected in degrees
 // Reference: https://en.wikipedia.org/wiki/Rotation_matrix
-void fill_rotation_matrix(rotation_vec3 rotation, double frequency, vec3 matrix[3]);
+void fill_rotation_matrix(rotation_vec3 rotation, float frequency, vec3 matrix[3]);
 
 // takes an old and new basis, and fills the passed vec3 array with the corresponding change basis matrix
 // allocate the passed vec3 array on the stack and pass the address to matrix
